+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Added support for query `PeerList`
 - Added support for query `PeerStats`
 - Added support for query `PeerVersion`
-- Added support for query `NodeInfo`.
+- Added support for query `NodeInfo`
+- Added support for query `StartBaker`
 
 ## 3.0.0
 - Stronger typing for various places in the API.